corna->conftests: improrve pgsql path comment
On a mac psql lives in different locations depending on how it was installed.

If installed via homebrew it goes `/usr/local/opts` if installed via the
official installer app from their website it goes to
`/Applications/Postgres.app/`.

This commit updates the comment to highlight this difference. The project
should decide on one official way to do it (probs homebrew) or add an extra
`if` statement to cover the extra case.