corna->Dockerfile-corna: move pip install above source copy
Currently, the last step in building the main corna docker image is to do a pip
install of the projects `requirements.txt`.

Do to he nature of software projects, typically source code moves a lot faster
than dependancies and requirements. As such it is better to place the slower
moving parts earlier in the image building so cache can be effectively used
making the image build faster.

This commit moves the `pip install` step before the source copy step in the
Dockerfile.