corna->utils: add ansible vault
This commit adds an ansible-vault to the Corna project to manage secrets.

Not this commit does not add ansible to the requirements, this is a stand alone
addition of an encrypted vault. The reason for this is the vault is due to reads
disproportionately out numbering writes and ansible is a huge requirement to add
to the project and slows down builds during the install step. As part of this
work we will add an inhouse implementation of reading values fro the vault
which will be significantly more lightweight than installing all of ansible just
read the values back out.

Having said that, adding/updating the vault will require a manual download of
`ansible===2.10.7`.