corna->utils: add vault crypto handling
Ansible is huge and we dont want to install it all just to be able to decrypt
the vault. This code has been lifted from the upstream ansible project for
dealing with the decryption elements of reading from the vault.

Note, to edit the vault itself we still need to install `ansible===2.10.7` as
this comes with the vault commands needed to do so. However, as this is not done
particularly often, it can be done on a case by case basis.