corna->vault_manager.py: add apis for interacting with the vault
This commit adds the main methods for getting both single and multiple values
out of the vault.

Note, due to decryption being an expensive operation we only do this once and
then cache the vault in memory. This means if the vault gets updated the entire
service running corna will need to restarted.