corna->env.py: add generated db url
Currently, alembic expects the database url (including password) in plain text
in the `alembic.ini` file. This is not secure and as the file will be checked
into the public repo adds extra issues.

This commit adds generation of url, similar to the db connection, so we do not
need to expose any sensitive information. We also override the config files
defaults used during the migrations/migration-file generation.

This does add the overhead of requiring all the necessary information for the
url to be generated (port, db-name, db-address etc).