corna->migrations: add first migration
This commit adds the first alembic migration file. It contains the migration
for the entire `models.py` in its current state as the migrations capability
had not been set up until now.

Obviously, this is not great and future migrations will be a part of the db
additions as and when they happen.