corna->secure.py: Decode messages
This commit adds a simple interface for decoding a message (currently a cookie)
from the incoming bytes to a string.

This makes any lookups in the database easier because we do out type
conversions in python without relying on postgres to do it for us. Type casts in
postgres have been a pain and quite buggy (mainly down to my own lack of
knowledge rather than anything to do with postgres itself).

This commit also adds the decode logic to anywhere that needs to do a db lookup
for a cookie.