corna->utils.py: Add correct type to mkdir signature
Previously, the `mkdir` function in utils arg type was `pathlib.Path` however,
the function actually does a type check and attempts to convert any string
into a path object before creating a directory.

This commit ensures this behaviour is reflected in the function signature by
giving the `path` arg a Union type of `pathlib.Path` and `str`.