corna->utils.py: Make `exists_ok` optional in `mkdir`
Previously, we suppressed all errors during `mkdir` as we did not care if the
directory existed or not.

For example, we are now using MD5 hashes as the base for image directory names,
it will be useful, if only for monitoring reasons right now, to know when/if
there is a hash collision (the file path will exist).

To allow us to optionally do this, we've parameterized `exists_ok` which is the
option in `pathlib.Path().mkdir` which deals with the check. If `exists_ok` is
false and there is a collision in directory names it will raise a
`FileExistsError` which we can anticipate for an log in our monitoring systems.