corna->utils.py: Update commend in `mkdir`
Previously, the comment above the actual call to `mkdir` said "recursively make
path", while this was not technically correct to begin with it was symbolic of
what happened in the call via the flags we passed in.

Now we have made it optional for `mkdir` to blow up, it was a good time to
change the comment to more accurately reflect what is happening, which is an
attempt at making the directory rather than a guarantee.