corna->utils.py: Add ability to get future date
We want to add expiry dates to cookies. In order to do that that we need to be
able to get a timezone aware date some time in the future.

This commit adds a function for doing that. The function also allows us to get
days in the past, meaning we can forcefully expire a session, if needed.