corna->secure.py: Add expiry date to message
Previously, we did not have a way to know when to expire a message, create a
limited lifetime message or force expire a message. The solution chosen to do
that is to simply add an expiry date to the message itself and use that as one
of the way to determine the validity of the message.

This commit adds an expiry date to signed messages.