corna->blueprints: gate create endpoints behind `login_required` flag
Previously, we did not check if a user was logged in or not before the main
body of the blueprint function.

This commit adds a quick check for login status before the request reaches the
main body, allowing us to abort early if a user is not logged in.

This gate has been added to all create endpoints (i.e. `POST`).