corna->themes: Add themes directory
This commit adds a theme directory to the root of the project. We decided
it should be at the root as we want clear separation between themes and
internal frontend code.

The eventual goal is that users can create themes for other users by
raising a PR in themes directory.