[bug] corna->image_proc.py: pass `FileStorage` into image hash
There was a bug which assumed that the image hash function was going to
receive a file path to a saved image but we are actually passing a `FileStorage`
object to the function.

This is because the hash is used to build the final destination of the image
and thus there is no file path to read from before the hash-ing has happened.

This commit fixes the bug.