[CI] corna: attempt to add mypy
Although we currently have some types being used, its mostly best
effort rather than being checked.

This commit adds mypy with very strict configuration. While there
has been some useful output, we will currently not be adding it to
CI. This is because, due to probably using an outdated version, there
are simply many places where the type checker is wrong.

It raises errors for `unexpected keyword argument` when they are very
much expected. It also raises other errors wuch as assignment errors
for things that are correctly typed e.g. enums.

While we wait to upgrade to a later version of mypy and gradully add
typing, we wont add it to the CI and continue with the best effort
approach currently being employed.

this commit adds mypy and tries to fix errors. Not all errors have been fixed and
this mypy has not been added to CI yet...