[CI] corna: use `ubuntu-20.04` for python3.6
Python3.6 is not supported on the `ubuntu-latest` image used in github actions.

This commit moves the image to a version which works.

Anoyingly, we also need to manually install and set up python3.6 on the image.
This commit also adds the steps for installing python3.6 as part of each CI
step.

Even more annoyingly, github actions does not really allow you to define a
shared step (easily) so for now we'll just copy and paste the build python3.6
step across each check.