corna->Makefile: break circular dependency
There is a weird circular dependency in our make file that causes the virtual
environment not to start correctly.

This commit comments out the break with the plan on fixing the issue at a later
date but in the (likely) case that it never happens, CI should not be affected.