corna->docker-compose.yml: Add simple health checks
Currently, there are no health checks on the containers that are created as
part of running the docker services for Corna.

This commit add simple health checks to the compose file that prevent the
services from running if both are not healthy. These checks are reoccurring
during the life cycle of the containers which makes them a good target for
monitoring using something like icinga (this can be done using docker inspect).

Note: for this to work properly `--wait` needs to be appended to the docker
up command so that the command does not exit before ensuring all services/
containers are healthy.