corna->docker-compose.yml: add vault password as a docker secret
Since the original creation of the docker compose config we have added a vault
to secure our secrets. For our code to access the vault it needs a password.

While simple, the difficulty with this approach is "How do you get the password
securely to the corna code?". A relatively straight forward approach to this
is using dockers secret function. This good thing about this approach is that
once the secret has been created by docker it doesn't need access to the original
file and thus it can be deleted from the host (what we are doing).

Contrary to documentation setting file permissions directly in the compose file
does not really seem to do anything, which is annoying. However, the file maintains
the permissions of the file on the host, we need to ensure we set those to 400
(or 600 whatever works).