corna->docker-compose.yml: always restart containers
Currently, we do not have a restart policy for our containers which default to
"restart: no". This is not ideal, we would rather they attempt to restart
unless explicitly removed, using docker compose.

This commit changes the restart policy to "always" meaning unless we run
`docker compose down` compose will try to restart the containers if they ever
stop running.