corna->.dockerignore: Add .dockerignore file
Surprisingly, up to the current point we have not had a .dockerignore file.

Our container for the main app are quite large (1.4gb), with the source copying
layer being surprsingly memory intensive (300mb). The hope is that by adding
files to ignore, we can begin slim down the image.

This commit adds a .dockerignore file which was mostly copied from github[1].

[1] https://gist.github.com/KernelA/04b4d7691f28e264f72e76cfd724d448