corna->docker-compose.yml: pull image from registry
Currently, we either have the docker image locally, which compose then runs
or we build the image from our local docker files. We now gave a docker registry
and would like to pull from the registry or build our image locally.
Adding the full registry URL to the image field satisfies this behaviour by
either pulling the image or creating it and tagging it with the URL/tag combo.

This commit adds the docker registry URL to the image fields of our docker
compose file.

More info on this behviour can be found here[1].

[1] https://github.com/docker/compose/issues/9730#issuecomment-1208342871