corna->Makefile: Remove use of venv in CI
Currently, CI is ran in a virtual environment, both locally and in GA's. We are
currently moving to a fully containerized local development environment which
emulates our production environment as closely as possible. As such, we want
to be able to run our make checks under the assumption that it will take place
in a container with all the necessary packages already installed, meaning we
do not need our Makefile to call any CI functions from a venv.

This commit removes references to `venv` in the Makefile and directly runs our
CI checks using `/usr/bin/python`.

We also update our GA's config to add a step which install's all the packages
in our `requirements.txt` file.