corna->Makefile: Build venv
Currently, we cannot really use `make` to build local `venv`'s because there
are some circular dependancies on `piptools` which essentially breaks the
command.

This commit untangles the circular dependencies and allows us to now use `make`
to both generate a new `requirements.txt` file and a full blow venv with all the
packages installed.