Corna->requirements.txt: bump versions of packages
In order to successfully get `make` to be able to create `venv`'s and generate
new requirements files we've done some local testing of the make commands which
gave resulted in us running `piptools`.

Running `piptools` shows there are slightly updated versions of our packages
which we can use instead of our current versions.

This commit adds the requirements file with bumped versions of our packages.