corna->Makefile: remove `pip==20.0.2` requirement
Currently, our services and `venv`'s require `pip==20.0.2` and it's not fully
clear why. Initially, it was believed to be due to that being a requirement
by `piptool` but after running `pip install --upgrade pip` and installing/running
`piptools` on a much later version of `pip`, everything seems to work fine.

We've also had issues installing ansible on `pip==20.*` but had none after
upgrading.

As a result of the above this commit removes the requirement of `pip==20.*` and
instead, upgrades `pip` to whatever the latest pip is (on `python3.6`).