corna->Dockerfle-nginx: Copy SSL certs
We now have SSL certs added to our vault meaning we can add them to the nginx
container.

Note: The temp files used to store the SSL keys are generated (and removed) as
part of our build process and do not actually exist locally, this is for
security reasons.

This commit adds the local location of our SSL certs in order to copy them into
our nginx docker container.