corna->Dockerfile-nginx: Copy frontend code into nginx
Currently, we use flask to find and return our static content to users. This
is very inefficient, especially when we are already using nginx as a reverse
proxy in our stack.

Instead of using flask we would like to take advantage of nginx and use it to
store and serve our static content to users.

This commit copies our frontend code into our nginx container and configures the
nginx file to server all static content (for the root url, not individual Corna
themes, we will do that separately).