ansible->build.sh: Simple build script
Currently, we have no automation for building and/or deploying our services to
a new container.

Although this is not exactly state of the art, it gives us a few things:
- Run CI on the current branch
- compile our typecript
- build and tag our service containers
- push new container to our remote registry

While there are a million ways to improve this (lots of those involve _not_ using
bash), this gives us something to get started with and a basis to improve from.