corna->compose.sh: Add docker compose wrapper script
This commit adds another simple (read: shitty) bash script. This script wraps
`docker compose` because now that we are passing all of our environment variable
to the compose file as _further_ commandline variables, it just makes the overhead
of managing to compose commands easier.

As a bonus we can use this script to run our compose file on our remote host.