corna->utils.py: Add project root
Our frontend (and soon themes) live as top level directories of the corna service.
In order to not always need to work it out, we've decided it would be useful
to have a constant variable pointing to the project root, relative to the
`utils.py` file.

This commit adds the constant variable which holds the path of the project root.