corna->meta.py: move `mkdir`
Currently, `mkdir` lives in `utils.py` as it is exclusively used by image
functionality in the `controls` (specifically post based controls).

However, as part of generalizing image processing functionality so it can be
used by other controls (in this case themes) `mkdir` will be used inside the
`image_proc` utility meaning we need to move it to `meta.py` to avoid any
circular dependencies.

This commit moves `mkdir` to `meta.py`.