corna->utils.py: move `random_short_string` to utils
Currently, the random short string generator lives in the posts control as that
is exclusively where it is used. However, as part of the work to generalize image
processing functionality to be used in other areas of the source code, other
controls will need access to the function.

This commit moves the `random_short_string` generator to `utils.py` because this
allows any are of our code to access the function.