corna->image_proc.py: move all image handling code
Currently, the only area of our code which needs to handle images is posts,
meaning that all image related code lives in `post_control`.

However, with more areas needing access to image processing functionality (e.g.
theme thumbnails) it makes sense to move the code that deals with this to the
`image_proc` util as it becomes accessible to all areas of our code.

This commit moves all necessary code into the `image_proc.py` util.