corna->utils.py: Centralize image validation
We are now at the point where we add/manage images in various different areas
across the code base, with even more to come.

This has resulted in a lot of repeated boiler plate code for verifying if an
uploaded file is in one of the accepted formats.

This commit centralizes all this code into one places and replaces any
pre-exiting checks to use the new code.