[Bug]corna->image_proc.py: Reset image seek pointer
We've seen during real testing that incoming image files do not get saved
correctly i.e. the file gets created but there are 0 bytes written to it.

After digging is seems that the initial `read()` we do when creating the images
MD5 hash causes the file reading pointer to read to the end of the file. Any
subsequent reads done on the file produce no bytes as the read pointer is
already at EOF.

In order to fix this we need to reset read pointer to the start of the file
after we compute the MD5 hash of the image.