corna->posts.py: Add HTML representation of post
Along with the core content of posts we want to save the HTML representation
of the post as its currently the only to display the post back out to the
browser.

This commit adds a new field to the database which holds the HTML and also
updates everything up the stack to ensure its correctly use.