corna->models.py: Add `orphaned` field to image table
We would like to add the notion that an image "may or may not have a parent post
or other associated data in the future" to the database.

There are cases that we may want to upload an image but cannot guarantee that it
will eventually be a part of a post or theme. For example, due to security
reasons browsers do not let your program (e.g. a rich text editor) access a
local file as the `src` attribute of an `img` tag. In order to get around this
we need to upload the image to the server first and then use its url in the
image tag we generate on the client.

However, this brings up the question, what if the user deletes the image while
editing the post and never actually uses it as part of the final post?

There are a couple approaches, delete image from server as the user deletes the
image during the editing process; leave the image as it and clean up as a
separate process.

With the first option, just watching the document for deleted/removed elements
is non-trivial. Then there is the incurred overhead of making a delete request
over the network, we already need to make two requests to upload the image and
then get it back again, so adding an unnecessary request also seems unwise.
Furthermore, what if the user acidentally deletes the image and then re-adds it?
Although, we could delay any deletes but because we do not know how long before
the delete the user will create the post, there is a risk of not deleting it
before the post is created thus, leaving an image that the user did not want
in the post.

The second option is simpler but requires us to keep track of the difference
between an image which is not a part of a post (e.g. a favicon) and an image
that will never be used because it was deleted from a post during editing.

This is where the `orphaned` field comes into play, any images saved during
editing will be marked as `orphaned`. We can then use this field in an automated
clean up process later on (i.e. delete any orphaned imaged created before some
date).

This commit adds an `orphaned` field to the images table to keep track of any
images that need to be cleaned up.