corna->pytest.ini: Add project `pyest.ini`
We would like to add a label to our tests which allows us to selectively not
use certain stubs/mocks which are set to `autouse=True` either at module level
or project level.

While it does not happen often, it can be a pain when you want to test a certain
testcase which requires an area of the code which has been mocked out. Rather
than manually adding all the mocks individually, because we still want them
mocked most of the time, the label will allow us to selectively "un-mock" any
testcases we do not need the project wide mocks for.

To do this we have added a `pytest.ini` to hold the created marks definition.
We've also filters out any `DeprecationWarning`'s as we know what they are and
will remove the filter when we come to update the project to a more modern
version of python.