corna->posts.py: Handle pre-loaded images
When creating a post in our editor on the users client there are cases where
the user uploads an image to the server as part of the editing process (we
upload the image to correctly show the preview in the `img` tag). This means
by the time the user as created the post we may already have several pieces
of orphaned artifacts in our db.

To fix this, we have added an extra field to the post payload which contains
a list of pre-existing images to link to the incoming post.

As part of this commit we also handle the actual linking process.