frontend->network.ts: Change `axios` header type
Currently, we use `AxiosHeaders` as the type for our headers but this type is
not particularly flexible and does not allow for you to add only select fields
to the header.

`RawAxiosRequestHeaders` does allow you to do this and is the base type for
axios headers.

This commit makes the change to using `RawAxiosRequestHeaders` instead of
`AxiosHeaders`.