corna->auth.py: Add domain name to cookie
Corna uses a lot subdomains, whether its for API requests or individual Corna's.
In order to share the session cookie across the base domain and all subdomains
we need to set the cookie domain to our base URL[1].

This commit adds the domain name to our cookie setting/removing code so it can
be handled properly in the browser. This should have no effect on direct API
use.

[1] https://stackoverflow.com/q/18492576