corna->image_proc.py: Dont save full image path
Currently, we save the full image path including the `PICTURE_DIR` which is
an environment variable which can change at any point. This means we can have
a full path which is actually invalid i.e. doesn't exit.

This commit changes the path we save from the full path (incl `PICTURE_DIR`) to
just saving the hash fs path which will always exist relative to whatever the
actual `PICTURE_DIR` is. Furthermore, its a good pattern for any future file
types we choose to save.