corna->image_proc.py: centralise file size calculator
Currently, we get the size of any files we save directly from the full path
that we've save it to. Now that we only save the hashed fs which does not know
about the `PICTURE_DIR` we need to access the `PICTURE_DIR` value every time
we want to get the size of a new file.

In order to simplify this process and keep it all in one place, we've added this
as separate functionality to `image_proc.py`.