corna->themes.py: use `login_required` instead of `before_request`
Currently, all theme endpoints require users to be logged in, this meant we could
use a global `login_required` gate before each request.  However, we soon
have some endpoints coming which do not need such a requirement.

This commit removes the global `before_request` check made for each request and
narrows the scope of the session check to each endpoint that needs it.