corna->auth.py: Delete pre-existing sessions
There is a bug that causes SQLAlchemy errors: client deletes cookie for
whatever reason e.g. browser was incognito and closed or cookies cleared, user
then tried to log back into the system but cannot due to the fact that they
already have a pre-existing session in the db against their name.

This commit ensures that any pre-existing sessions get removed when a user is
attempting to log in. This essentially guarantees a clean slate whenever a
user attempts to login.