corna->posts.py: Remove images from text post
Forms do not do lists well *sigh*.

Currently, our text-post endpoint expects a list of uuids associated with any
pre-uploaded inline images that are a part of the core content. This is the
main way we currently allow users to add images to a text post on the client.

We are sending the data as a form and sending one value is fine, but the moment you try send more than one value one of two things can happen:
  - only the last item added gets sent
  - it sends all the values with the same key

This is a pain to deal with. There was an attempt to just stringify the list on the client but because marshmallow expects a form, it did not deserialize the list properly, breaking the saving process.

The solution? Just remove the whole form thing. Most of our data for each post
is strings so we can just use JSON. If there are any files clients want to post they can us the media upload endpoint separately. This removes all the headaces.

This commit removes the form requirements from the text-post upload endpoint and changes it to JSON.