docker->corna: Copy only necessary files
Currently, our docker build command copies over all source files in our main
directory (i.e. the build context), this is not great because at build time
there are various secrets generated e.g. SSL keys and lots of other fluff we
do not want as part of the container.

This commit copies over only the relevant and needed files. This is probably a
placeholder as in the future, this is not how we want to work with static files
such as those in the frontend and themes directories.