docker-compose.yml: `PICTURE_DIR` should map to directory inside container
Currently, we use an environment variable for the picture directory location
which maps to the directory where the images _come from_ on the host machine.
This is different to the location the file will get mounted on the container.

This commit fixes this by ensuring that the `PICTURE_DIR` envvar that gets
passed to the actual service is pointing to the right place.