corna->posts.py: catch `PostDoesNotExistError`
When we first introduced linking pre-existing images to an incoming post we
decided to raise a `PostDoesNotExistError` if the pre-existing image is not
found. However, we do not catch this exception on the API response meaning it
can silently fail and not tell the client anything meaningful.

This commit ensures that the API response catches the error and handles it
appropriately.