setup.cfg: ignore E712
Annoyingly, you can't ignore/disable pycodestyle errors for a particular line.
In sqlalchemy comparision checks of a boolean i.e. `foo == True` is actually
a meaningful thing because of how the SQL that is generated.

In order for this error to not be raised as a result of using it in sqlalchemy
queries we choose to ignore it. Its worth noting that `pylint` also raises the
same error so we will not be missing out on this check going forward.

[1] https://github.com/PyCQA/pycodestyle/issues/480