Dockerfile-nginx: move the copy of `nginx.conf`
The nginx.conf file moves a lot faster than the TLS certs i.e. has more changes.
As such it makes more sense to do the copy step of the conf file after the certs
as this will make building the containers smoother/quicker.

This commit moves the `nginx.conf` copy step to happen after the TLS certs are
copied.