nginx.conf: set max body size and max buffer size
We've seen issues where nginx fails to pass through large images/files because
it does not have enough buffer space. Locally, this is usually due to not
allocating docker enough memory.

However, we still do not want the risk of files not saving due to a small and
obscure error. This commit sets a relatively high buffer size for nginx.