[refactor] remove **kwargs to make function argument clearer
Currently, we pass argument parameters to most of our control code as dicts.
This is a problem because is obfuscates what the actual params to the function
are and it also means we always have to use TypedDict's to create the type
signatures.

This commit "unpacks" the dicts with the function arguments by making the
function signatures more explicit.

This commit also does some slight renaming of parameters to avoid naming collisions.