corna->alchemy.py: add wrapper around core sqlalchemy actions
There is a lot of repeated code over the repo for largely doing a lot of the
same actions e.g. get a row from a table in the DB.

This commit adds a simple wrapper around those actions (as of this commit, not
all actions have a wrapper) which allows us to not only centralize changed to
the ORM but also allows use to handle errors better as we currently do no
error handling on any of our DB requests/queries.