corna->corna.py: Add base permissions for a corna
While we currently have the ability to give individual users roles we want
to also allow users to set base permissions that affect every user that comes
to that corna.

For example, if a user wants to make a private corna it is
impractical (and impossible) to give every visiting user a role which says that
the user _cannot_ read anything on the corna. In cases like this it makes much
more sense to make the whole corna private and then give individual uses read
access.