corna->post_control.py: Save post creator
Currently, our post infrastructure has always assumed only the user of a given
corna would be the one making posts, thus not saving the creator information.

However, we now allow non-owner users to create posts on any corna they have the permissions to do so. As a result we now need to differentiate between the each
user who created the post, meaning saving the posters information is valuable.

This commit adds the above functionality.