tests/tests_media.py: skip webp image test
It seems that support for detecting the mimetype of webp file formats was
added to pythons mimetype module in the python3.11 release [1], as result we
skip the webp media test.

The skipping is done via a conditional so will automatically run once we update
to python3.11 without having to change anything.

We also separate gif tests from webp tests by adding a new test that uses a
a gif with a `.gif` extension instead of the previous `.webp` extension.

We also add a few extra comments around using a more robust file type detection
method (like libmagic), there are some links for reading when the time comes to
fix this, its not very pressing atm.

[1] https://github.com/python/cpython/pull/29259#pullrequestreview-960709458