core: Upgrade to python3.12
This commit upgrades the corna python version to 3.12. We also update our
packages to reflect this change.

Our current version of SQLAchemy (1.4+) requires us to set environment variables
which silence deprecation warnings. These warnings are useful as part of upgrading
from 1.* > 2.* but that is outside the scope of this upgrade.