core: use local copy of flask-sqlalchemy-session
flask-sqlalchemy-session's latest release is broken by Werkzeug 2.1 removing
__ident_func__ from flask._app_ctx_stack. A PR fixing this deprecation has been
raised however support for the package seems to have been abandoned given this
fix for a breaking change has not been merged despite being raised 7 months ago:
https://github.com/dtheodor/flask-sqlalchemy-session/pull/15

To circumvent this, this commit creates a local copy of the PRs code to be used
directly and removes the pip import for flask-sqlalchemy-session.

A new directory, 'oss', has been created to house this and similar copies of
external code. The Makefile has been updated to ignore linting here on the
basis that this code may not always match our linters but should be preserved
in its original form.