corna->auth_control.py: ensure cookie is a string before setting
Currently, we try to set the user cookie with a byte string, not a python string.
This has caused a bug as flask's set_cookie function expects the cookie to be
a string.

In the earlier version of flask we were using (this commit is part of the
python3.6 > 3.12 upgrade), flask did not seem to complain about this so the
bug went unnoticed.

This commit fixes the bug by ensuring that the signed cookie is converted into
a python string before returning to the caller.