corna->models.py: Ensure password SHA string field is the right length
The hashing function used by werkzeug is SHA256 and thus returns 256 bytes for
every password.

Currently, our password string field is expecting a 128 byte string. Although,
this was a known bug, for some reason, until we upgraded our version of python
(this commit is part of the python3.6 -> python3.12 upgrade), SQLAlchemy did
not complain about this (maybe it was silently truncating the pwd *shrug*).

This commit fixes the password length field the the correct length of 256.