pytest.ini: only ignore flask_apispec deprecation warning
Currently, we ignore all deprecation warnings as we know we are significantly
behind in terms of python versions.

As we are now upgrading to python3.12 we no longer need global warnings.

However, it seems that the package `flask_apispec` has not updated marshmallow
versions, which is causing deprecation warnings. There is a issues[1] along with
a PR bit its been open for several months and does not look like it will be
merged soon.

In the meantime, we want to ignore `flask_apispec` specific deprecation warnings
as its a 3rd party package and we have no control over whether it will be updated
or not.

[1] https://github.com/jmcarp/flask-apispec/issues/255