corna->workflows: modernize GHA CI pipelines
Now that we have updated from python3.6->python3.12 we can start using modern
tools like everyone else.

We no longer need to:
  * manually install python - we can use a premade GHA python setup step
  * use an old version of ubuntu - we can now move to using `ubuntu-latest`

We've also made a few enhancements like using the GHA matrix strategy to test
our code on multiple versions of python (this is only used for tests).

A final key change this commit makes is directly invoking the CI using
`python` rather than `make`. This is because our makefile expects the
python binary to be located in `/usr/bin/python` but GHA places the
binary in an arbitrary location. This means that CI is currently out of
lockstep with our make file, going forward the solution will likely
be to simply use the python keyword in our makefile but that will
come later as we will uplift our dockerfile away from CentOS 7.