corna->media_control.py: directly use the mimetypes lib
We are still seeing problems with the `webp` tests in the media lib
as we're currently relying on werkzurg to be using the latest Cpython
updates and cannot be sure that they are.

It seems that Cpython has added the `webp` mimtype to the lastest version
of python[1] so we can directly use it ourselves.

Although, this library is still fundamentally flawed as it relies on the
file extension, it seems to do the job for now.

As such this commit does not remove the improvement comment inlined in
the code as we still probably need to move to a `libmagic` based library
which will determin the mimetype based on the file contents rather than
file extension.

[1] https://github.com/python/cpython/blob/3.12/Lib/mimetypes.py#L598