dockerfiles/Dockerfile-corna: update from CentOS 7 -> Rocky 9
This commit moves us off CentOS and onto Rocky which is the CentOS replacement
after RedHat fucked shit up.

We don't use a minor version as this allows us to take advantage of Rockys rolling
updates, as of this commit we are on Rocky 9.4. The rolling updates change the
minor version of each release.