dockerfiles/Dockerfile-corna: locale setting fix
For some reason the `en_GB.utf8` file is not a standard part of rocky9 which
means we have to manually install the language packs to be able to set the
container locale.

Furthermore, `yum-config-maneger` is also not on  rocky9 by default however,
the `yum` command can directly set and save the installation locale.

This commit fixes both issues.