dockerfiles/Dockerfile-corna: install python3.12
This commit installs python3.12 into the container. By default, rocky 9 comes
with python3.9 preinstalled.

Rocky 9.4 (which is the current latest minor version) adds python3.12 as an RPM
and because we are using the generic `rockylinux:9` tag for our container we are
automatically updated to the latest rocky minor release (currently 9.4) meaning
we are be able to install python3.12 using dnf/yum.

NOTE: if for whatever reason we need to install rocky 9.3 or lower, this step
will not work.