compose.sh: create a tmp copy of SSL certs
Previously, we were not able to build the nginx container while doing local
development as we didn't have functionality for adding the SSL certs to the
container (or we simply commented them out in the container).

This commit adds a temp cert creation step which gets the cert from the vault
all copies it to a local file for the duration of the build set.