dockerfile: set owner/perms for requirements file
When docker copies a file, it does so with the same permissions as the host
machine. More often than not, this means the owner has root permissions.

The user in our container does not have root perms so we need to set allow the
user to have read access to the requirements file before installing them.

This commit adds ownership settings to the requirements file during the copy step.